etl/requirements.txt
flake8==4.0
python-dotenv==0.21.1
psycopg2-binary==2.9.5
python-dotenv==0.21.1
pyyaml==6.0

elasticsearch~=7.17.9
pydantic~=1.10.5

